fix - bug no cadastro de produto

+++ src/main/java/ZtechAplication/repository/MarcaRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.repository.CrudRepository;
 import ZtechAplication.model.Categoria;
 import ZtechAplication.model.Marca;
 
-public interface MarcaRepository extends CrudRepository<Marca, Long> {
+public interface MarcaRepository extends CrudRepository<Marca, Integer> {
 	Optional<Marca> findByNome(String cate);
 
 }
